+++ common/src/main/java/tocraft/walkers/integrations/impl/AlexMobsIntegration.java
@@ -21,22 +21,22 @@ public class AlexMobsIntegration extends AbstractIntegration {
     @Override
     public void registerTraits() {
         // PreyTrait
-        registerAlexMobsTargetTagTrait("anaconda");
-        registerAlexMobsTargetTagTrait("bald_eagle");
-        registerAlexMobsTargetTagTrait("cachalot_whale");
-        registerAlexMobsTargetTagTrait("crimson-mosquito");
-        registerAlexMobsTargetTagTrait("crocodile");
-        TraitRegistry.registerByType(getType("fly"), (PreyTrait<Mob>) PreyTrait.ofHunterTag(getTag("fly_hurt_targets")));
-        registerAlexMobsTargetTagTrait("froststalker");
-        registerAlexMobsTargetTagTrait("giant-squid");
-        registerAlexMobsTargetTagTrait("komodo-dragon");
-        registerAlexMobsTargetTagTrait("mantis_shrimp");
-        registerAlexMobsTargetTagTrait("orca");
-        registerAlexMobsTargetTagTrait("snow_leopard");
-        registerAlexMobsTargetTagTrait("sunbird_scorch");
-        registerAlexMobsTargetTagTrait("tiger");
-        registerAlexMobsTargetTagTrait("warped_toad");
-        TraitRegistry.registerByType(getType("bunfungus"), new PreyTrait<>(List.of(entity -> (entity instanceof Enemy && !(entity instanceof Creeper) && !(entity.getType().getCategory().name().contains("water") && entity.isInWaterOrBubble()) && !entity.getType().is(getTag("bunfungus_ignores"))))));
+        registerAlexMobsTargetTagTrait("anaconda", 1, 200);
+        registerAlexMobsTargetTagTrait("bald_eagle", 4, 55);
+        registerAlexMobsTargetTagTrait("cachalot_whale", 2, 30);
+        registerAlexMobsTargetTagTrait("crimson-mosquito",2, 50);
+        registerAlexMobsTargetTagTrait("crocodile", 5, 180);
+        TraitRegistry.registerByType(getType("fly"), new PreyTrait<>(List.of(), List.of(), List.of(), List.of(getTag("fly_hurt_targets")), 3, 5));
+        registerAlexMobsTargetTagTrait("froststalker", 3, 80);
+        registerAlexMobsTargetTagTrait("giant-squid", 3, 70);
+        registerAlexMobsTargetTagTrait("komodo-dragon", 8, 180);
+        registerAlexMobsTargetTagTrait("mantis_shrimp", 3, 120);
+        registerAlexMobsTargetTagTrait("orca",3 , 200);
+        registerAlexMobsTargetTagTrait("snow_leopard", 2, 10);
+        registerAlexMobsTargetTagTrait("sunbird_scorch", 3, 5);
+        registerAlexMobsTargetTagTrait("tiger",4 ,220);
+        registerAlexMobsTargetTagTrait("warped_toad",4, 50);
+        TraitRegistry.registerByPredicate(entity -> (entity instanceof Enemy && !(entity instanceof Creeper) && !(entity.getType().getCategory().name().contains("water") && entity.isInWaterOrBubble()) && !entity.getType().is(getTag("bunfungus_ignores"))), new PreyTrait<>(List.of(), List.of(getType("bunfungus")), List.of(), List.of(), 3, 5));
         // FearedTrait
         registerAlexMobsFearTagTrait("mimic_octopus");
         registerAlexMobsFearTagTrait("skunk");