+++ CHANGELOG.md
@@ -4,6 +4,7 @@ walkers 5.1
 - fix blacklist commands stack
 - add FearedTrait & PreyTrait for Alex's Mobs
 - fix ConcurrentModificationException on AbilityRegistry
+- rework registration of PreyTrait & FearedTrait
 
 walkers 5
 ================